Divide pipeline items into their own files, add functionality
Adds rss crawl compare for postgres databases, which was previously only
implemented for mysql. This is not a good solution, since most of the code
is copied from the other class. This should be changed in the future by
implementing database as repositories and decoupling the rss crawler from
the implementation of the database. Rss Crawl Compare Postgres is implemented
as a scrapy middleware, instead of a pipeline element for news-please. This
makes sure that the page is not downloaded if the url is already present
in the database, which speeds up the process.

This commit also changes the DontProxyMiddleware class to work with the
generic proxy Middleware of scrapy

Also fixes some minor problems with the BeautifulSoup Parser